+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - "Strings" search mode for strings on one line starting and ending with `"`
 - "Comments" search mode for comments on one line starting with `#` or `//`
+- "Ls" search mode to display statistics about directory tree
 
 ## [1.0.0] 2021-05-18
 ### Added